Background story and UI manager added
Added a background story, a typewriter effect was enabled for the background story.
An enum was created to record the scene state. Now there are two scene state:
1. Background Story
2. Game Start
3. Pick Gear
Created a UI manager. Will use the manager to control the placement of different UI components and the switch between different scene state.