Add SelectionPanelController, GearManager, GearShowcase, Updated GearData and GearDatabase
Added a new class GearDataInGame within the GearData scriptableobject script. This class is created because the Gear data will be changed in game, but the scriptable object is designed to remain the same during different game runs. So this class is used and store in the InGame Deck and will be changed(Upgrade or change the features of it) during the game run.

Added some new method in GearDatabase, now it is able to  generate random gear with desired amount and rarelevels from the data list.

Created a gearManager to store and manages the gearDataInGame which is the gears that players picks.

Created a prefab for gearCard

Created SelectionPanelController, control the state of UI components in the selection panel. Now it will be called in the select gear scene, and the gear card in the panel will be automatically generated from the GearCard prefab.

Created a GearShowcase for GearCard, it will show all kinds of GearInfo on the gear card.

Plan:
Added Hover function for the gearShowcase.
Added selection method for the gearcard in the gearShowcase Script
Update GearManager when select gears(Added Gear)
Update SelectionPanelController so that more/less options will be supported(Now it is fixed to 3 option)